# Optimise Rust build